botrite/requirements.txt
asgiref==3.6.0
backports.zoneinfo;python_version<"3.9"
dj-database-url==1.2.0
Django==4.1.6
gunicorn==20.1.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
twilio==7.12.1
pandas==1.5.3
openpyxl==3.1.1
djangorestframework==3.14.0
django-pygments==0.5
pygments==2.14.0
celery>=5.0.5
geopy>=2.4.1
shapely
django-wsgi
